move files to their respective component folder
