+++ src/components/result/result.tsx
